Add chain-of-trust validation
This also fixes a bug in the PKCS#7 Signature object parser: one cannot
assume that the 'certificates' entry in SignedData is ordered, since it
is defined as a set in ASN.1. In particular, this means that we have to
look for the signer cert by cross-referencing with the 'sid' entry in
SignerInfo.