Add functionality to encapsulate external pages
The import_page_as_xobject function grabs a page content stream from
another PdfHandler object, and imports it as a form XObject into the
current one (including all relevant resources). The idea is that this
might be useful to reuse pages from an external PDF as images in another
one.