Advanced modification detection
 - First step: implement more intelligent signature coverage checking
 - I also threw out the embedding metadata tracking in the reader,
since it creates more issues than it solves.
 - Evaluate modification level by iterating over xrefs in incremental
   update

The current implementation is not terribly efficient, and does not yet
support the finer points described in the standard.
In particular, the results of the analysis aren't yet compared against
/DocMPD permissions, and the validator is very conservative about
allowing field modifications (basically, it assumes that all fields
with a /V entry are always locked)