Deal with nested fields in field enumeration
The current implementation ignored fully qualified names and did not
resolve /FT correctly. It would also produce invalid files when
presented with a field name containing a period, so I put a safeguard in
place to prevent that until we have support for creating fields at
deeper locations in the form hierarchy.