Integrate new diff analysis with validation API
The validation endpoints now accept an optional diff policy as
input, and can request that the diff analysis be skipped.

The PdfSignatureStatus now has an extra diff_result field,
and the modification_level property is computed based on
the value of diff_result in combination with the signature's coverage
level.