Fixed another Xref stream handling bug
The compression parameters of previous Xref streams weren't treated any
differently as trailer values by the updating code. This caused them to
clobber the parameters for future Xref streams as well. If both streams
were compressed using the exact same settings, this isn't an issue, but
in other cases it does matter.

Additionally, the Xref /W and /Index would get copied from the old
stream, which is also undesirable.

(this came up when finicking around with updates on a file written by
Acrobat)