Minor I/O tweaks
Most changes are in the signing routines:

 - Chunk reads when reading input from a file-like object (as opposed to
an in-memory buffer or mmap'd file)
 - Allow specifying the output stream/object via output=...

This commit also adds a convenience method to IncrementalPdfFileWriter
to output only changed objects. This was possible before, but required
a call to a piece of private API.

Note: output= has limited usefulness, since the output stream must support
reads as well. This is an unfortunate corollary of the way pyHanko computes
hashes.

An alternative approach would be to support outputting only the changed
objects after signing (should be small enough to fit in memory), but that
creates issues when PDF files are read/modified repeatedly (PAdES
operations). I'll be shelving those optimisations for the time being.