Work around an issue in oscrypto
oscrypto's OpenSSL-based implementation of aes_cbc_no_padding_encrypt
requires padding to encrypt a single 16-byte block if the key
size is larger than 16 bytes. I believe that's a mistake, but since
I'm not willing to vendor oscrypto, I put in a little hack to work
around that in the routine that encrypts the permission bits for
PDF 2.0 encryption.