diff_analysis: multiply referenced objects
The difference analysis tool is now path-aware
(at a fairly hefty performance cost).
mkl convinced me that this was necessary to correctly evaluate some
particularly hairy decisions involving modifications to the structure
tree of a document. Right now, pyHanko is blissfully unaware of any tag
structure modifications regardless, but it's something we might want to
support in the future.

Regardless, not allowing a multiply-referenced object to be overridden
_by default_ is a sane thing to do anyway, and these changes
should make writing safe rules somewhat easier.

More polishing is needed, but tests seem to work OK.