Allow key usage to be specified in the config file
This commit allows CLI users to override the default key usage
requirements for signing certificates by means of the new
'signer-key-usage' and 'signer-extd-key-usage' keys in a validation
context. This configuration option does not have a corresponding command
line flag (yet).

These values also override the default nonRepudiation check performed in
the signing code (so users who explicitly want to sign with a
certificate that doesn't have the nonRepudiation bit can do so).

See discussion in #10.