Populate /AP on empty visible signature field
This commit simply inserts a light grey rectangle with a narrow black
border.

Note: appearances for empty signature fields are ignored or changed
in most viewers that actually support signatures, but this is not a big
deal. This change still ensures that the empty signature field renders
in less feature-rich viewers, and also satisfies the requirement
mandating that all visible annotations have an appearance stream.