Tweak interpretation of skip_diff
Now, skip_diff causes all changes to be accepted in the validation
report, instead of (effectively) rejecting everything. This allows
skip_diff to be used as a quick workaround for cases where the
incremental update analysis is overly zealous.

The latter is arguably better addressed by the 'no changes' diff policy.