Add variable text keys to diff whitelist
This commit whitelists the keys /DA and /Q for use in form field
dictionaries and the global /AcroForm dictionary.

Acrobat will occasionally set these values if not present, regardless of
whether the file is signed or not. Since these keys are generally fairly
harmless, we ignore updates to them during incremental update analysis.