Allow SKI to identify signer certificates
This crosses off a small issue that had been on the backlog for quite a
while. Note: many PDF viewers/SDKs apparently don't support this
functionality. This might have to do with the fact that it didn't exist
in OG PKCS#7, but that shouldn't stop us from supporting it in our
validation workflow.