Complete overhaul of text rendering & layout
 - Use HarfBuzz for shaping (incl. complex scripts)
 - Support non-CFF TrueType fonts
 - Support vertical writing
 - Use ActualText marked content in addition to ToUnicode
 - Introduce simple box layout & alignment rules, and
   apply them uniformly across all layout decisions where possible
 - Simplify code in text box / stamp layouting classes accordingly

Note that the text layout system is still relatively dumb: it's only
capable of shaping (through HarfBuzz) with a single font.
It doesn't do any text segmentation, bidi processing, styling, ...

In particular, combining different fonts or scripts is still not
possible, but this update represents a large step in the right
direction.

Note: the API for layout / text processing is still largely internal,
and subject to change in the future!