Properly handle escapes in name objects
This fixes another holdover from PyPDF2. Note: name objects are still
treated as UTF-8 strings internally, while the standard technically
doesn't specify a character encoding (so they have to be treated as raw
bytes that happen to map to a 7-bit ASCII encoding in the end).

This is not an issue since the overwhelming majority of PDF names expand
to ASCII text, and >99.9% of those are even alphanumeric. On the other
hand, correcting this typing issue in the object model we inherited from
PyPDF2 would cause massive API breakage both within and outside of
pyHanko, so I'll leave that as it is for now.