Align CertificateStore API with pyHanko
SimpleCertificateStore was originally just a shim in pyHanko to avoid
having to fork certvalidator, but that ship has sailed, so now that need
the same functionality in pyhanko_certivalidator either way, it's time
to ditch those classes from pyHanko once pyhanko_certvalidator is
released.

This commit rectifies a number of outstanding post-refactor issues to
make that transition possible without breaking existing code.