Pass stream_xrefs=False in copy_into_new_writer()
This function transparently ignores object streams anyhow, so the xref
stream is pointless in most cases (unless the caller wants to put in
their own object streams).

The right thing to do is to make stream_xrefs=True the default,
expose it as a function argument _and_ handle object streams more
cleverly when constructing the new writer. That optimisation will have
to wait for now.