Remove obsolete/misguided md_algorithm fallback
Falling back to the DigestMethod entry on author signatures was a
mistake, caused by a misguided TODO I added in the very early days (see
commit eb0613e), which I implemented rather uncritically in
e15dcff.

There's no reason to believe that the DigestMethod in a sigref dictionary
means anything remotely useful in this day and age. Given that some
libraries hardcode it to designate horribly outdated hash functions,
this fallback was even potentially harmful. Hence why I'm removing it.