Fix confusing fetch behaviour in PKCS#11 signer
Instead of refusing to pull certs when the 'ca_chain' parameter is
anything other than None, just determine whether fetching is necessary
in the method that pulls the certs, and return early it if fetching is
unnecessary.

This behaviour was the underlying cause of a bug in the PKCS#11 signing
CLI, since the config module always passes an empty iterable into
'ca_chain'. The effect was that no matter what, it was impossible to get
pyHanko to fetch intermediate certs from the PKCS#11 token when invoked
from the CLI.

Exception: this doesn't apply to the BeID signer, since that one is
configured differently.