Avoid unguarded next() in _walk_page_tree_annots()
LGTM complained about using unguarded next()s in generator code,
which is indeed an antipattern (due to the overloaded meaning of
StopIteration in earlier Python versions), but in this case we
can actually guarantee that it won't trigger.
Nonetheless, better to change it to something that won't throw
StopIteration no matter what.