No longer allow bytes instead of x509.Certificate
Previously, many entry points in the public API would try to be smart
about automatically decoding byte strings passed as certs. This isn't
really tenable if we want to support attribute certificate validation,
so the feature was removed. All certificates must now be passed in
as asn1crypto.x509.Certificate objects.

This is a breaking API change, but the real-world impact is probably
limited, especially since pyHanko offers scaffolding to deal with this
stuff out of the box.