Fix critical bug in content timestamp generation
The document digest was being re-hashed. This went unnoticed because the
validation code only took the content timestamp check into account on raw CMS
signatures (by accident). This was fixed as a side effect of 2b45006,
thus causing the bug to surface.

This commit fixes the bug in question and improves the relevant tests:

 - The existing "happy path" test was improved to test the integrity
   more thoroughly.
 - Added an additional test with a content timestamp purposely generated
   based on a wrong digest.