Further regression tests for content timestamp bug
See bug fixed in 853c5a2.

As noted in the commit message for 853c5a2, the content timestamp status
was effectively only checked in CMS-only signatures, but due to typos in
a test case there, the bug went undetected there as well (ugh...).

This commit fixes the affected test case and adds another regression
test on the CMS side, by analogy with the tests added in 853c5a2.