Control hybrid xref support with strict mode
Instead of using a separate flag when signing/validating, we now use the
strict mode flag on the underlying reader to determine whether it's OK
to use hybrid references or not.

This commit also adds a flag to IncrementalPdfFileWriter to control
whether to read the underlying file in strict mode.

See #71.