Start moving towards non-cert trust roots
This means: breaking changes in the ValidationPath class, which I have
tried to limit to modifying APIs that wouldn't be that useful outside
this package.

Note that the actual main validation loop still can't deal with non-cert
trust roots, but that's mostly a matter of initialisation of the PKIX
parameters, which I intend to (largely) subsume into the TrustAnchor class
either way.

Currently all tests in pyhanko-certvalidator pass, but the refactoring is
not complete yet, and lacks tests for the new functionality. Still a
WIP.