Clean up some meta stuff from parent repo
Among other things, I removed the `docs/api.md` file. The idea is to
replace the old `certvalidator` with RST documentation that can be
integrated into the main pyhanko API reference. But since the current
Markdown API doc file was outdated and no longer updatable, I figured it
would be better to remove it altogether for the time being. The pyhanko
docs have enough examples to cover basic use cases.