Fix typing issue in read_dss()
See #81.

The default for an absent entry in the DSS would be an empty tuple,
which is problematic for later modification. Hence, when updating the
DSS in a document with (say) /OCSPs and /Certs in the DSS but not /CRLs,
the addition of any new CRLs would result in an AttributeError when
attempting to call `append()` on the tuple.

This commit fixes #81 and adds a regression test.