Initial work towards ser/deser of file credentials
This is the first step towards resolving #91: we need a way to persist
file access info in a clean way. Even if not persisting any key
material, there should be a mechanism in place to automatically
reconstruct access from said serialised data, with the same access level
(insofar as that matters).

Steps taken in this commit:

 - Basic, mixin-friendly API setup
 - Implementation for the standard security handler