Add experimental "fancy QR code" option
i.e. support for QR codes with graphics in the middle (+ a higher error
correction rate to compensate, and more stylish rounded modules)

The refactoring of the QR-generating code also triggered some very minor
changes to the ways the relevant QR codes are laid out in the PDF file.
Those changes all have to do with coordinate system transformations, and
the difference is visually almost imperceptible---they are probably due
to simple rounding issues on the renderer's end. As such, the relevant
layout test files have been replaced with regenerated ones.