Preparatory refactor of trust config code in CLI
Process configuration in multiple phases (first generating policies, and
then the validation context) instead of spitting out a validation
context from the start. This is needed to support AdES validation in the
CLI because the AdES layer doesn't take external validation context objects.

This also changes the API of CLIConfig in a breaking way, removing the
get_validation_context() method.